% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genome_Info_Mining.R
\name{nsrgenomes}
\alias{nsrgenomes}
\title{nsrgenomes}
\usage{
nsrgenomes(species_name, accession_number)
}
\arguments{
\item{species_name}{The name of the organsim of interest. Ex. "E_coli"}

\item{accession_number.}{The accession number of the organsim of interest. separate multiple accesion numbers by comma with no spaces}
}
\value{
mRNA and rRNA/tRNA .csv files
}
\description{
get the mRNA and rRNA/tRNA sequences for an organism of interest from the NCBI database
}
\examples{
nsrgenomes("S_mutans","NC_004350.2")
}
